[fix]Fix warning by static_cast.

+++ cocos/base/ccUtils.cpp
@@ -131,7 +131,7 @@ void captureNode(Node* startNode, std::function<void(RefPtr<Image>)> imageCallba
         
         RenderTexture* finalRtx = nullptr;
         
-        auto rtx = RenderTexture::create(size.width, size.height, backend::PixelFormat::RGBA8, PixelFormat::D24S8);
+        auto rtx = RenderTexture::create(static_cast<int>(size.width), static_cast<int>(size.height), backend::PixelFormat::RGBA8, PixelFormat::D24S8);
         // rtx->setKeepMatrix(true);
         Point savedPos = startNode->getPosition();
         Point anchor;
@@ -152,7 +152,7 @@ void captureNode(Node* startNode, std::function<void(RefPtr<Image>)> imageCallba
             Sprite *sprite = Sprite::createWithTexture(rtx->getSprite()->getTexture(), finalRect);
             sprite->setAnchorPoint(Point(0, 0));
             sprite->setFlippedY(true);
-            finalRtx = RenderTexture::create(size.width * scale, size.height * scale, backend::PixelFormat::RGBA8, PixelFormat::D24S8);
+            finalRtx = RenderTexture::create(static_cast<int>(size.width * scale), static_cast<int>(size.height * scale), backend::PixelFormat::RGBA8, PixelFormat::D24S8);
             
             sprite->setScale(scale); // or use finalRtx->setKeepMatrix(true);
             finalRtx->begin();
